lab 4/5 fix

+++ CGLab5/shader.frag
@@ -46,9 +46,9 @@ void main()
 		//рассеяная составляющая
 		vec3 L = light.position - fragCoord;
 		vec3 N = normal;
-		float cosLN = dot(L, N) / (length(L) * length(N));
+		float cosLN = float(dot(L, N) / (length(L) * length(N)));
 		vec3 I_d = material.k_d * light.intensity * max(0, cosLN);
-		I_d /= pow(length(L), 2) * light.attenuation;
+		I_d /= float(pow(length(L), 2) * light.attenuation);
 		//отраженная составляющая
 		vec3 I_s;
 		if (cosLN > 0){
@@ -56,9 +56,9 @@ void main()
 			vec3 normalizedL = normalize(L);
 			vec3 R = ((cosLN * N) - normalizedL) + N * cosLN;
 			vec3 S = normalize(cameraDirection);
-			float cosRSp = pow(max(0, dot(R, S) / (length(R) * length(S))), material.p);
+			float cosRSp = float(pow(max(0, dot(R, S) / (length(R) * length(S))), material.p));
 			I_s = material.k_s * light.intensity * cosRSp;
-			I_s /= pow(length(L), 2) * light.attenuation;
+			I_s /= float(pow(length(L), 2) * light.attenuation);
 		} else {
 			I_s = vec3(0, 0, 0);
 		}
